requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
boto3==1.40.6
twilio==9.3.2
mangum==0.19.0
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.4.2
python-dotenv==1.0.1
pytest==8.3.4
